+++ src/main/java/lotto/controller/Controller.java
@@ -24,7 +24,7 @@ public class Controller {
 
             //2로또 랜덤 번호 생성
             outputView.showCount(count);
-            GenerateLottoNumbers generateLottoNumbers = generateNumbers(count);//로또 랜덤 번호 생성 및 검증 진행
+            GenerateLottoNumbers generateLottoNumbers = generateNumbers(count);//로또 랜덤 번호 생성 후 로또 클래스에서 검증 진행
             // 랜덤으로 생성된 로또번호 보여주기
             outputView.showLottos(generateLottoNumbers);
