+++ src/main/java/lotto/domain/Lotto.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
 import java.util.List;
-import lotto.validator.oldLottoValidator;
+import lotto.validator.LottoValidator;
 
 public class Lotto {
     private final List<Integer> numbers;
@@ -9,8 +9,8 @@ public class Lotto {
     // 이 로또 클래스를 어떻게 써야좋을까
 
     public Lotto(List<Integer> numbers) {
-        oldLottoValidator.validate(numbers); // ✅ 여기서 검증
-        this.numbers = List.copyOf(numbers);
+        LottoValidator.validate(numbers); // ✅ 여기서 검증
+        this.numbers = List.copyOf(numbers); // 불변 리스트
     }
